Extract generic stream encoder adaptor
