+++ src/stream/generic/mod.rs
@@ -1,0 +1,3 @@
+mod encoder;
+
+pub(crate) use encoder::Encoder;
